pyproject.toml
[tool.poetry]
name = "exchange_rates"
version = "0.1.0"
description = ""
authors = ["Leonid Lygin <ionagamed@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
django = "^3.0"
djangorestframework = "^3.11"
psycopg2-binary = "^2.8"
flake8 = "^3.7"
django-cacheops = "^4.2"
requests = "^2.23"
drf-yasg = "^1.17"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
